backend/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
defusedxml==0.8.0rc2
Django==5.0.6
django-cors-headers==4.4.0
django-filter==24.2
django-oauth-toolkit==2.4.0
django-rest-framework-social-oauth2==1.2.0
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
dnspython==2.6.1
drf-yasg==1.21.7
email_validator==2.2.0
exceptiongroup==1.2.1
h11==0.14.0
httpcore==1.0.5
httpx==0.25.2
idna==3.7
inflection==0.5.1
install==1.3.5
jwcrypto==1.5.6
oauthlib==3.2.2
packaging==24.1
pillow==10.4.0
pycparser==2.22
pydantic==2.8.0
pydantic_core==2.20.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
pyuploadcare==5.1.0
PyYAML==6.0.1
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
sniffio==1.3.1
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
